requirements.txt
alembic==1.7.1
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
colorama==0.4.4
coverage==5.5
cryptography==3.4.8
decorator==5.0.9
dnspython==2.1.0
flasgger==0.9.5
Flask==1.1.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mistune==0.8.4
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.23
validators==0.18.2
waitress==2.0.0
Werkzeug==1.0.1
